--- AulaController.java
@@ -1,7 +1,0 @@
-
-package com.ska.NEXUS.controllers;
-
-
-public class AulaController {
-    
-}
