--- CadastroController.java
@@ -1,6 +1,0 @@
-
-package com.ska.NEXUS.controllers;
-
-public class CadastroController {
-    
-}
