--- CadastroService.java
@@ -1,6 +1,0 @@
-
-package com.ska.NEXUS.service;
-
-public class CadastroService {
-    
-}
