--- DespesaModel.java
@@ -1,53 +1,0 @@
-package com.ska.NEXUS.model;
-
-
-
-public class DespesaModel {
-
-    private int id;
-    private String descricao;
-    private float valor;
-    private String data;
-    private String categoria;
-
-    public int getId() {
-        return id;
-    }
-
-    public String getDescricao() {
-        return descricao;
-    }
-
-    public float getValor() {
-        return valor;
-    }
-
-    public String getData() {
-        return data;
-    }
-
-    public String getCategoria() {
-        return categoria;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public void setDescricao(String descricao) {
-        this.descricao = descricao;
-    }
-
-    public void setValor(float valor) {
-        this.valor = valor;
-    }
-
-    public void setData(String data) {
-        this.data = data;
-    }
-
-    public void setCategoria(String categoria) {
-        this.categoria = categoria;
-    }
-
-}
