--- EmailService.java
@@ -1,5 +1,0 @@
-package com.ska.NEXUS.service;
-
-public class EmailService {
-    
-}
