--- UsuarioDTO.java
@@ -1,53 +1,0 @@
-package com.ska.NEXUS.dto;
-
-public class UsuarioDTO {
-
-    private Long id;
-
-    private String nome;
-
-    private String senha;
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    private String usuario;
-    private String email;
-
-    public String getUsuario() {
-        return usuario;
-    }
-
-    public void setUsuario(String teste) {
-        this.usuario = teste;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String senha) {
-        this.senha = senha;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
