--- UsuarioRepositoryCustom.java
@@ -1,11 +1,0 @@
-
-
-package com.ska.NEXUS.repository.custom;
-
-import com.ska.NEXUS.dto.UsuarioDTO;
-
-
-public interface UsuarioRepositoryCustom {
-    public boolean inserirUsuario(UsuarioDTO usuarioDTO);
-    
-}
