--- autenticacaoDTO.java
@@ -1,23 +1,0 @@
-package com.ska.NEXUS.dto;
-
-public class autenticacaoDTO {
-
-    private String usuario;
-    private String senha;
-
-    public String getUsuario() {
-        return usuario;
-    }
-
-    public void setUsuario(String teste) {
-        this.usuario = teste;
-    }
-
-    public String getSenha() {
-        return senha;
-    }
-
-    public void setSenha(String oi) {
-        this.senha = oi;
-    }
-}
