+++ back-end/src/main/java/com/ska/NEXUS/controllers/UsuarioController.java
@@ -1,28 +1,59 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package com.ska.NEXUS.controllers;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
+import com.ska.NEXUS.dto.UsuarioDTO;
+import com.ska.NEXUS.dto.autenticacaoDTO;
+import com.ska.NEXUS.service.AutenticacaoService;
+import com.ska.NEXUS.service.UsuarioService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/api")
-public class UsuarioController{
+@RequestMapping("/usuarioController")
+@CrossOrigin(origins = "*")
+public class UsuarioController {
 
-@GetMapping("mensagem")
-public String mensagem(@RequestParam String texto){
-    return "mensagem recebida: " + texto;
-}
+   
+    @PostMapping
+    @RequestMapping("/autenticar")
+    public ResponseEntity<?> autenticar(@RequestBody autenticacaoDTO dto) {
+        AutenticacaoService service  = new AutenticacaoService();
+        
+        // ? = qualquer classe
+        //ResponseEntity<autenticacaoDTO>
+        
+        boolean autenticado = service.autenticar(dto);
 
+         System.err.println(autenticado);
+        
+        if (autenticado) {
+            return ResponseEntity.ok("Usuário autenticado com sucesso!");
+        } else {
+            return ResponseEntity.status(401).body("Usuário ou senha inválidos!");
+        }
+    }
+    
+    @GetMapping
+    @RequestMapping("/teste")
+    public String autenticar(@RequestParam String dto) {
+        return "teste " + dto;
+    }
+    
+    
+    @PostMapping
+    @RequestMapping("/cadastro")
+    public ResponseEntity<?> cadastro(@RequestBody UsuarioDTO dto) {
+        UsuarioService service  = new UsuarioService();
+        
+        boolean cadastrado= service.cadastro(dto);
 
-@GetMapping("mensagem2")
-public String mensagem2(){
-return "OLA SOU UMA MENSAGEM!";
-}
-
-
+         System.err.println(cadastrado);
+        
+        if (cadastrado) {
+            return ResponseEntity.ok("Usuário autenticado com sucesso!");
+        } else {
+            return ResponseEntity.status(401).body("Usuário ou senha inválidos!");
+        }
+    }
+    
+    
 }
