+++ back-end/src/main/java/com/ska/NEXUS/dto/autenticacaoDTO.java
@@ -1,0 +1,23 @@
+package com.ska.NEXUS.dto;
+
+public class autenticacaoDTO {
+
+    private String usuario;
+    private String senha;
+
+    public String getUsuario() {
+        return usuario;
+    }
+
+    public void setUsuario(String teste) {
+        this.usuario = teste;
+    }
+
+    public String getSenha() {
+        return senha;
+    }
+
+    public void setSenha(String oi) {
+        this.senha = oi;
+    }
+}
