+++ back-end/src/main/java/com/ska/NEXUS/repository/custom/UsuarioRepositoryCustom.java
@@ -1,0 +1,11 @@
+
+
+package com.ska.NEXUS.repository.custom;
+
+import com.ska.NEXUS.dto.UsuarioDTO;
+
+
+public interface UsuarioRepositoryCustom {
+    public boolean inserirUsuario(UsuarioDTO usuarioDTO);
+    
+}
